Add thread allocation helpers
